backend/services/service-transcriptions/requirements.txt
quart
motor
pillow
python-dotenv
pydantic
pydash
pymongo
PyMuPDF
pytesseract
pytest
pytest-asyncio
pytest-mock